feat(article):重构文章和评论实体及接口，支持点赞、收藏与浏览记录功能
- 修改 Article 和 ArticleComment 实体类字段类型，统一使用 Timestamp 表示时间
- 新增 ArticleUser 和 ArticleCommentUser 实体及相关 Mapper、Service，用于管理用户对文章和评论的点赞、收藏等状态
- 更新 ArticleController 和 ArticleCommentController，增加推荐文章、获取个人文章、点赞评论等功能接口
- 调整 MyBatis 查询逻辑，完善文章和评论的关联查询，支持返回资源文件和用户交互状态（如是否点赞）
-优化参数校验和空值判断，提升接口健壮性和 Swagger 文档可读性

+++ studyParty_entityApi/src/main/java/com/studyParty/entity/user/DTO/UserDTO.java
@@ -24,7 +24,7 @@ public class UserDTO implements Serializable {
     private String grade;
     private String remark;
     private int status;
-    private int starPrestige;
+    private Long starPrestige;
     private String phone;
     private String school;
     private int clockIn;
